+++ CHANGELOG.md
@@ -14,8 +14,8 @@ Versioning].
 
 ### Security
 
-- Upgrade Go to `v1.26.1`.
-- Upgrade go-git from to 5.17.1.
+- Upgrade Go to `v1.26.2`.
+- Upgrade go-git to `v5.17.1`.
 
 ## [v0.6.3] - 2026-01-07