+++ CHANGELOG.md
@@ -14,7 +14,7 @@ Versioning].
 
 ### Security
 
-- Upgrade Go to `v1.26.3`.
+- Upgrade Go to `v1.26.4`.
 - Upgrade go-git to `v5.19.1`.
 
 ## [v0.6.3] - 2026-01-07